refactor: extract broker settings view models
